my_app/backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
alembic==1.14.1
argcomplete==1.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.3.0
blinker==1.8.2
breezy==3.2.1
catkin-pkg-modules==1.0.0
ccimport==0.4.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.16.1
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
contourpy==1.2.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cumm-cu102==0.4.11
cycler==0.12.1
Cython==0.29.37
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
debugpy==1.8.1
Deprecated==1.2.13
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docutils==0.17.1
docx==0.2.4
dulwich==0.20.31
einops==0.8.0
empy==3.3.4
exceptiongroup==1.2.1
executing==2.0.1
fastbencode==0.0.5
fastimport==0.9.14
fastjsonschema==2.19.1
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.47.2
fsspec==2024.5.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
gpg==1.16.0
greenlet==3.1.1
grpcio==1.64.1
gyp==0.1
h5py==3.11.0
hdbscan==0.8.37
httplib2==0.20.2
idna==3.3
importlib_metadata==7.1.0
iniconfig==1.1.1
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keras==3.4.1
keyring==23.5.0
kiwisolver==1.4.5
labelCloud==1.1.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mmcv==2.2.0
mmengine==0.10.4
more-itertools==8.10.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
nose2==0.9.2
notify2==0.3
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
open3d==0.18.0
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
optree==0.11.0
optuna==4.2.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
patiencediff==0.2.1
pccm==0.4.11
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==0.13.0
pointnet==0.1.1
portalocker==2.10.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pybind11==2.13.1
pycairo==1.20.1
PyGithub==1.55
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.7.2
PyNaCl==1.5.0
pyntcloud==0.3.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquaternion==0.9.9
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-gitlab==2.10.1
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.4
rich==13.7.1
roman==3.3
rpds-py==0.18.1
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.15.1
SecretStorage==3.3.1
serial==0.0.97
six==1.16.0
spconv-cu102==2.3.6
SQLAlchemy==2.0.38
stack-data==0.6.3
systemd-python==234
taichi==1.7.1
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminator==2.1.3
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
visidata==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.13.3
yapf==0.40.2
zipp==1.0.0
